+++ README.md
@@ -1,7 +1,10 @@
-# poker
-Oppgave 54 fra ProjectEuler.com
+# Poker
+I don't really know anything about poker but have made a hold'em poker variation. At the moment this is purely a terminal game.
 
--Oppdatering
-Jeg misforsto oppgaven, gjør den rikitg nå i poker_task.py. 
-Tenker å lage den første main mer interaktiv.
+# Functionality
+* Two players currently (adding query for more inn soon)
+* Stats, win rate, wins, losses and ties if that somehow occurs
+* Clean UI in terminal
+* Easy quitting
 
+Originally this was just me doing task 54 from ProjectEuler.com but i misunderstood(i did not see the input file) and made a poker game instead.
